exp now reduces all types by a % value

+++ The Demon King/Assets/Scripts/EvolutionSystem/ExperienceBar.cs
@@ -14,7 +14,7 @@ public class ExperienceBar
 
 
 
-    [HideInInspector] public int CurrentExp;
+    [HideInInspector] public float CurrentExp;
 
     [Header("EXPERIENCE REQUIRED TO EVOLVE")]
     public IntSO level1ExpNeeded;
